crawler/requirements.txt
# 웹 크롤링 관련
requests>=2.31.0
beautifulsoup4>=4.12.3
selenium>=4.18.1
webdriver-manager>=4.0.1

# Supabase 연동
supabase>=2.3.4

# 이미지 처리
Pillow>=10.2.0

# 환경변수 관리
python-dotenv>=1.0.1

# 유틸리티
python-dateutil>=2.8.2
